st3: install autowrap (again), setup lineendings
- tell lineendings to scream when someone's careless text editor usage
  on windows butchers the line endings.
- give autowrap another chance, this time with
  "auto_wrap_break_long_word" set to false (obviously still happens in
git comments, but that's a whole nother issue. basically, if the hard
wrap is enforced even in a markdown file, if the markdown renderer
renders markdown correctly, the paragraph will not actually be broken if
there's a linebreak in the file between two words - that is, unless
there are *two spaces* between the words. best of both worlds there,
because then the unrendered markdown is readable and not just a bunch of
incredibly long lines. kthxbai